refactored - extracted `<GameConfigDialog/>`
